artifacts/code/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "factory-control-bot"
version = "1.0.0"
description = "Telegram bot for Black Box Software Factory v2 command and control"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Factory Team"},
]

dependencies = [
    "python-telegram-bot[all]>=21.5,<22",
    "sqlalchemy[asyncio]>=2.0.30,<2.1",
    "asyncpg>=0.29.0,<0.30",
    "alembic>=1.13.0,<1.14",
    "pydantic>=2.7.0,<3",
    "pydantic-settings>=2.3.0,<3",
    "aiodocker>=0.22.0,<0.23",
    "httpx>=0.27.0,<0.28",
    "psutil>=5.9.0,<6",
    "structlog>=24.1.0,<25",
    "python-dotenv>=1.0.0,<2",
    "aiofiles>=23.2.0,<24",
    "orjson>=3.10.0,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<0.24",
    "pytest-cov>=5.0,<6",
    "testcontainers[postgres]>=4.4,<5",
    "ruff>=0.4.0,<1",
    "mypy>=1.10,<2",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests (require DB)",
    "e2e: End-to-end tests",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
